Unify Apple folder/address lists onto the sidebar
Bring the mailbox sidebar lists to full parity with the old management
views, then retire the management views so there is one list per data
type on the Apple clients.

- Port createFolder/deleteFolder/possibleParents/canDelete onto
  FolderListViewModel; add a "New folder" toolbar button + sheet and a
  gated Delete swipe/context action + confirmation to FolderListView.
- Add a "Request new address" toolbar button + NewAddressSheet and a
  Revoke swipe/context action + confirmation to AddressListView (the
  shared AddressesViewModel already had revoke/create hooks).
- Extract NewFolderSheet into its own file for reuse.
- Compact iPhone keeps its Addresses/Folders tabs, now backed by the
  unified list views via thin NavigationStack wrappers.
- Reduce the macOS Settings window and the iPad-regular settings sheet
  to General only; folder/address admin now lives in the sidebar.
- Delete FoldersAdminView, AddressesView, FoldersAdminViewModel, and the
  now-unused macOS RequiresSignIn; refresh stale doc comments.
