+++ changelog.d/apple-unify-folder-address-lists.changed.md
@@ -1,0 +1,4 @@
+- Apple clients: folder and address management is now built into the mailbox
+  sidebar lists. Creating and deleting folders and requesting and revoking
+  addresses happen right where you browse them, and the separate management
+  screens have been removed.
